fix: handle optional env and break cache module cycle
