+++ package.json
@@ -17,7 +17,7 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "export $(cat .env.test | xargs) && NODE_ENV=test && jest --config ./test/jest-e2e.json"
+    "test:e2e": "bash -c 'if [ -f .env.test ]; then export $(cat .env.test | xargs); fi; NODE_ENV=test jest --config ./test/jest-e2e.json'"
   },
   "dependencies": {
     "@google-cloud/bigquery": "^7.9.1",
