+++ src/cache/cache.constants.ts
@@ -1,0 +1,8 @@
+export interface CacheConfig {
+  redisUrl?: string;
+  maxObjectBytes: number;
+  gcsBucket?: string;
+}
+
+export const CACHE_CONFIG = 'CACHE_CONFIG';
+export const GCS_BUCKET = 'GCS_BUCKET';
